add story for Transcription Editor
